pcsx2: Update to v1.7.4823-pre recommended nightly
This build of pcsx2 is now built against Qt6 and the binary is named pcsx2-qt.

Upstream is Windows oriented, so this package is installed to /usr/share/pcsx2/bin/
which is somewhat unconventional (to put it mildly).

We add pcsx2 and pcsx2-qt compatibility symlinks to a /usr/bin/pcsx2-qt.sh shell script
as a convenience for existing v1.6.0 users.

Additionally, the .desktop application launcher is patched to start from the correct path.

Thanks to @ZachBacon for the work he did re. getting the package installable.

Test Plan:

Install and run SSX Tricky with the Vulkan renderer at 3x internal resolution.

Generally speaking, on the same hardware, this version is more performant and smoother
when compared to the old 1.6.0 release we used to have in the repo. It also sports a
Vulkan backend which worked wonderfully during testing.

Signed-off-by: Rune Morling <ermo@serpentos.com>